commit23(part15)

+++ frontend/src/components/WorkoutDetails.js
@@ -1,11 +1,12 @@
 import { useState } from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
-
+import { useAuthContext } from '../hooks/useAuthContext'
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   // State for editing
   const [isEditing, setIsEditing] = useState(false)
@@ -14,8 +15,14 @@ const WorkoutDetails = ({ workout }) => {
   const [editReps, setEditReps] = useState(workout.reps)
 
   const handleClick = async () => {
+    if (!user) {
+      return
+    }
+
+
     const response = await fetch(`${process.env.REACT_APP_API_URL}/api/workouts/${workout._id}`, {
       method: 'DELETE',
+      headers: {'Authorization': `Bearer ${user.token}`}
     })
     const json = await response.json()
